Simplified structure TimeslotSet. Added compile-time choice for TimeslotSet data structure, which may be represented by a bool array of a long integer (as bit string). Added candidatesCount to State, which is updated when candidate timeslots are discarded from TimeslotSet.
Changed selectTimeslot: now the timeslot chosen is the one which the least number of other unplaced events could be assigned, and if more than one, choose the timeslot with the most events in, ties are broken randomly.

When all rooms of a timeslot are assigned to events, this timeslot is now discarded from candidates of all unplaced events.

The state stack is now dynamically allocated. The Solution instance is now updated when a feasible solution is found.